seperate models implementation
