+++ src/app/core/models/feels-like.model.ts
@@ -1,0 +1,6 @@
+export interface FeelsLike {
+    day: number;
+    night: number;
+    eve: number;
+    morn: number;
+  }
