+++ src/app/core/models/openweather-api.model.ts
@@ -1,3 +1,6 @@
+import { Current } from "./current.model";
+import { Daily } from "./daily.model";
+
 export interface OpenWeatherApiResponse {
   name: string;
   image: string;
@@ -9,73 +12,3 @@ export interface OpenWeatherApiResponse {
   hourly: Current[];
   daily: Daily[];
 }
-
-export interface Current {
-  dt: number;
-  sunrise?: number;
-  sunset?: number;
-  temp: number;
-  feels_like: number;
-  pressure: number;
-  humidity: number;
-  dew_point: number;
-  uvi: number;
-  clouds: number;
-  visibility: number;
-  wind_speed: number;
-  wind_deg: number;
-  weather: Weather[];
-  snow?: Snow;
-  pop?: number;
-  rain?: Rain;
-}
-
-export interface Snow {
-  '1h': number;
-}
-export interface Rain {
-  '1h': number;
-}
-
-export interface Weather {
-  id: number;
-  main: string;
-  description: string;
-  icon: string;
-}
-
-export interface Daily {
-  dt: number;
-  sunrise: number;
-  sunset: number;
-  temp: Temp;
-  feels_like: FeelsLike;
-  pressure: number;
-  humidity: number;
-  dew_point: number;
-  wind_speed: number;
-  wind_deg: number;
-  weather: Weather[];
-  clouds: number;
-  pop: number;
-  rain?: number;
-  snow?: number;
-  uvi: number;
-}
-
-export interface FeelsLike {
-  day: number;
-  night: number;
-  eve: number;
-  morn: number;
-}
-
-export interface Temp {
-  day: number;
-  min: number;
-  max: number;
-  night: number;
-  eve: number;
-  morn: number;
-}
-
