+++ src/app/core/models/rain.model.ts
@@ -1,0 +1,3 @@
+export interface Rain {
+    '1h': number;
+  }
