+++ src/app/core/models/snow.model.ts
@@ -1,0 +1,3 @@
+export interface Snow {
+    '1h': number;
+  }
