+++ src/app/core/models/temp.model.ts
@@ -1,0 +1,8 @@
+export interface Temp {
+    day: number;
+    min: number;
+    max: number;
+    night: number;
+    eve: number;
+    morn: number;
+  }
