+++ src/app/core/models/weather.model.ts
@@ -1,0 +1,7 @@
+export interface Weather {
+    id: number;
+    main: string;
+    description: string;
+    icon: string;
+  }
+
